# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Communication_LTD Ticket System API
  description: |
    API for Communication_LTD customer support ticket system
    
    ## Logging Requirements
    
    - All successful mutating operations (POST, PUT, PATCH, DELETE) with valid CSRF tokens are logged at INFO level
    - Missing or invalid CSRF tokens are logged at WARNING level
    - Requests approaching rate limit thresholds are logged at WARNING level
    - All requests include client IP, user ID (if authenticated), timestamp, and operation details
    
    ## Rate Limiting
    
    The API implements rate limiting to protect against abuse:
    - Authentication endpoints: 10 requests per minute per IP
    - Standard endpoints: 60 requests per minute per authenticated user
    - Admin endpoints: 30 requests per minute per admin user
    
    When requests approach 80% of their limit threshold, WARNING level logs are generated
  version: 1.0.0
  contact:
    name: Communication_LTD Support
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      description: >
        CSRF token for mutation requests protection.
        All POST, PUT, PATCH, DELETE operations require this header.
        Missing or invalid tokens will be logged at WARNING level.
        Valid tokens on successful mutating requests will be logged at INFO level.
      required: true
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      required:
        - code
        - message
      example:
        code: 400
        message: Bad Request
        details:
          field: "username"
          reason: "Username is already taken"
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customer_id:
          type: integer
          format: int64
        assigned_to:
          type: integer
          format: int64
          nullable: true
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [new, in_progress, on_hold, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        category:
          type: string
          enum: [billing, technical, account, other]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - customer_id
        - subject
        - description
        - status
        - priority
        - category
        - created_at
        - updated_at
      example:
        id: 1
        customer_id: 1
        assigned_to: 2
        subject: "Internet connection issue"
        description: "My internet connection has been unstable for the past 2 days."
        status: "new"
        priority: "high"
        category: "technical"
        created_at: "2025-08-10T14:30:00Z"
        updated_at: "2025-08-10T14:30:00Z"
        closed_at: null
    TicketCreate:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        category:
          type: string
          enum: [billing, technical, account, other]
      required:
        - subject
        - description
        - category
      example:
        subject: "Internet connection issue"
        description: "My internet connection has been unstable for the past 2 days."
        priority: "high"
        category: "technical"
    TicketUpdate:
      type: object
      properties:
        assigned_to:
          type: integer
          format: int64
          nullable: true
        status:
          type: string
          enum: [new, in_progress, on_hold, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
      example:
        assigned_to: 2
        status: "in_progress"
        priority: "high"
    TicketComment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ticket_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - ticket_id
        - user_id
        - content
        - created_at
        - updated_at
      example:
        id: 1
        ticket_id: 1
        user_id: 2
        content: "We're looking into this issue. Please provide more details about when the problem occurs."
        created_at: "2025-08-10T15:00:00Z"
        updated_at: "2025-08-10T15:00:00Z"
    TicketCommentCreate:
      type: object
      properties:
        content:
          type: string
      required:
        - content
      example:
        content: "We're looking into this issue. Please provide more details about when the problem occurs."
...
paths:
...
  /tickets:
    get:
      summary: List tickets
      operationId: listTickets
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - name: status
          in: query
          schema:
            type: string
            enum: [new, in_progress, on_hold, resolved, closed]
          description: Filter by ticket status
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
          description: Filter by ticket priority
        - name: category
          in: query
          schema:
            type: string
            enum: [billing, technical, account, other]
          description: Filter by ticket category
        - name: customer_id
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by customer ID
        - name: assigned_to
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by assigned user ID
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer
                  limit:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new ticket
      operationId: createTicket
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        '201':
          description: Ticket created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tickets/{ticket_id}:
    parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get ticket details
      operationId: getTicket
      tags:
        - Tickets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update ticket details
      operationId: updateTicket
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdate'
      responses:
        '200':
          description: Ticket updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a ticket
      operationId: deleteTicket
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Ticket deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tickets/{ticket_id}/comments:
    parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get ticket comments
      operationId: getTicketComments
      tags:
        - Tickets
        - Comments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of ticket comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketComment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Add comment to ticket
      operationId: addTicketComment
      tags:
        - Tickets
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCommentCreate'
      responses:
        '201':
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketComment'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...